Refactor text tokenizers
