+++ src/fairseq2/data/text/__init__.py
@@ -8,12 +8,18 @@ from fairseq2.data.text.converters import StrSplitter as StrSplitter
 from fairseq2.data.text.converters import StrToIntConverter as StrToIntConverter
 from fairseq2.data.text.converters import StrToTensorConverter as StrToTensorConverter
 from fairseq2.data.text.sentencepiece import (
+    BasicSentencePieceTokenizer as BasicSentencePieceTokenizer,
+)
+from fairseq2.data.text.sentencepiece import (
     SentencePieceDecoder as SentencePieceDecoder,
 )
 from fairseq2.data.text.sentencepiece import (
     SentencePieceEncoder as SentencePieceEncoder,
 )
 from fairseq2.data.text.sentencepiece import SentencePieceModel as SentencePieceModel
+from fairseq2.data.text.sentencepiece import (
+    SentencePieceTokenizerBase as SentencePieceTokenizerBase,
+)
 from fairseq2.data.text.sentencepiece import (
     vocab_info_from_sentencepiece as vocab_info_from_sentencepiece,
 )
