+++ src/fairseq2/models/nllb/tokenizer.py
@@ -6,24 +6,15 @@
 
 from typing import Optional, Sequence, Set, final
 
-from fairseq2.data.text import (
-    SentencePieceDecoder,
-    SentencePieceEncoder,
-    SentencePieceModel,
-    TextTokenDecoder,
-    TextTokenEncoder,
-    TextTokenizer,
-    vocab_info_from_sentencepiece,
-)
+from fairseq2.data.text import SentencePieceEncoder, SentencePieceTokenizerBase
 from fairseq2.data.typing import PathLike
 from fairseq2.typing import Device, finaloverride
 
 
 @final
-class NllbTokenizer(TextTokenizer):
+class NllbTokenizer(SentencePieceTokenizerBase):
     """Represents the tokenizer used by NLLB models."""
 
-    model: SentencePieceModel
     langs: Set[str]
     default_lang: str
 
@@ -49,15 +40,11 @@ class NllbTokenizer(TextTokenizer):
         # it to the model at index 0.
         control_symbols.append("<pad>@0")
 
-        self.model = SentencePieceModel(pathname, control_symbols)
+        super().__init__(pathname, control_symbols)
 
         self.langs = set(langs)
 
         self.default_lang = default_lang
-
-        vocab_info = vocab_info_from_sentencepiece(self.model)
-
-        super().__init__(vocab_info)
 
     @finaloverride
     def create_encoder(
@@ -68,7 +55,7 @@ class NllbTokenizer(TextTokenizer):
         mode: Optional[str] = None,
         device: Optional[Device] = None,
         pin_memory: bool = False,
-    ) -> TextTokenEncoder:
+    ) -> SentencePieceEncoder:
         """Create a token encoder.
 
         :param task:
@@ -127,13 +114,3 @@ class NllbTokenizer(TextTokenizer):
             device=device,
             pin_memory=pin_memory,
         )
-
-    @finaloverride
-    def create_raw_encoder(
-        self, *, device: Optional[Device] = None, pin_memory: bool = False
-    ) -> TextTokenEncoder:
-        return SentencePieceEncoder(self.model, device=device, pin_memory=pin_memory)
-
-    @finaloverride
-    def create_decoder(self) -> TextTokenDecoder:
-        return SentencePieceDecoder(self.model)
